+++ lib/extract_messages.dart
@@ -24,6 +24,12 @@ import 'dart:io';
 import 'package:analyzer/analyzer.dart';
 import 'package:intl/src/intl_message.dart';
 
+/// A function that takes a message and does something useful with it.
+typedef void OnMessage(String message);
+
+/// What to do when a message is encountered, defaults to [print].
+OnMessage onMessage = print;
+
 /// If this is true, print warnings for skipped messages. Otherwise, warnings
 /// are suppressed.
 bool suppressWarnings = false;
@@ -57,8 +63,8 @@ Map<String, MainMessage> parseFile(File file, [transformer = false]) {
   try {
     root = parseDartFile(file.path);
   } on AnalyzerErrorGroup catch (e) {
-    print("Error in parsing ${file.path}, no messages extracted.");
-    print("  $e");
+    onMessage("Error in parsing ${file.path}, no messages extracted.");
+    onMessage("  $e");
     return {};
   }
   origin = file.path;
@@ -190,8 +196,9 @@ class MessageFindingVisitor extends GeneralizingAstVisitor {
         var err = new StringBuffer()
           ..write("Skipping invalid Intl.message invocation\n    <$node>\n")
           ..writeAll(["    reason: $reason\n", _reportErrorLocation(node)]);
-        warnings.add(err.toString());
-        print(err);
+        var errString = err.toString();
+        warnings.add(errString);
+        onMessage(errString);
       }
       // We found one, but it's not valid. Stop recursing.
       return true;
@@ -266,8 +273,9 @@ class MessageFindingVisitor extends GeneralizingAstVisitor {
         var err = new StringBuffer()
           ..writeAll(["Error ", e, "\nProcessing <", node, ">\n"])
           ..write(_reportErrorLocation(node));
-        print(err);
-        warnings.add(err.toString());
+        var errString = err.toString();
+        onMessage(errString);
+        warnings.add(errString);
       }
       return message; // Because we may have set it to null on an error.
     }
@@ -448,8 +456,9 @@ class PluralAndGenderVisitor extends SimpleAstVisitor {
         var err = new StringBuffer()
           ..writeAll(["Error ", e, "\nProcessing <", node, ">"])
           ..write(_reportErrorLocation(node));
-        print(err);
-        warnings.add(err.toString());
+        var errString = err.toString();
+        onMessage(errString);
+        warnings.add(errString);
       }
     });
     var mainArg = node.argumentList.arguments
@@ -464,8 +473,9 @@ class PluralAndGenderVisitor extends SimpleAstVisitor {
             "must be simple variable reference) "
             "\nProcessing <$node>")
         ..write(_reportErrorLocation(node));
-      print(err);
-      warnings.add(err.toString());
+      var errString = err.toString();
+      onMessage(errString);
+      warnings.add(errString);
     }
     return message;
   }